ai_arena/generator/requirements.txt
psycopg2-binary==2.8.6
psycopg2==2.8.4
requests==2.27.1
requests_toolbelt==0.9.1
google-cloud-pubsub==2.12.0
google-cloud-tasks==1.5.0
google-cloud-logging==3.0.0
protobuf==3.20.0